Refine squad table design, columns, and sorting
Clean up the squad overview table and make its columns behave consistently
across view modes:

- Identity: replace the colored avatar circle with a plain muted squad
  number; split it into its own sortable "#" column (name + nationality
  stay together).
- Quieter repeated values: morale shows a colored dot only (label on hover),
  energy bar is slimmer, position chips are left-aligned.
- Anchor the common columns (#, player, position, age, average) at fixed
  widths so they no longer shift when switching tabs; trailing spacer absorbs
  slack. Widen morale/contract/age so headers no longer truncate.
- Planning tab: surface the next-season projection inside the potential bar
  and drop the now-redundant dev-status column.
- Add an up/down "sortable" affordance icon to every header (directional
  arrow when active).
- Sort POS by canonical position order (PositionMapper::positionSortOrder),
  GK -> DEF -> MID -> FWD, instead of alphabetically by abbreviation.
- Extract the inline Alpine component to resources/js/squad-overview.js.
