+++ resources/js/app.js
@@ -8,6 +8,7 @@ import lineupManager from './lineup';
 import matchSummaryLineups from './match-summary-lineups';
 import negotiationChat from './negotiation-chat';
 import squadSelection from './squad-selection';
+import squadOverview from './squad-overview';
 import tournamentSummary from './tournament-summary';
 import seasonSummary from './season-summary';
 import squadRegistration from './squad-registration';
@@ -27,6 +28,7 @@ Alpine.data('lineupManager', lineupManager);
 Alpine.data('matchSummaryLineups', matchSummaryLineups);
 Alpine.data('negotiationChat', negotiationChat);
 Alpine.data('squadSelection', squadSelection);
+Alpine.data('squadOverview', squadOverview);
 Alpine.data('squadRegistration', squadRegistration);
 Alpine.data('tournamentSummary', tournamentSummary);
 Alpine.data('seasonSummary', seasonSummary);
