+++ resources/views/squad.blade.php
@@ -14,95 +14,7 @@
         <x-game-header :game="$game" :next-match="$game->next_match"></x-game-header>
     </x-slot>
 
-    <div x-data="{
-        viewMode: new URLSearchParams(window.location.search).get('mode') || 'tactical',
-        posFilter: new URLSearchParams(window.location.search).get('pos') || 'all',
-        availFilter: new URLSearchParams(window.location.search).get('avail') || 'all',
-        statusFilter: 'all',
-        sortCol: new URLSearchParams(window.location.search).get('sort') || null,
-        sortDir: new URLSearchParams(window.location.search).get('dir') || 'desc',
-        sidebarOpen: true,
-
-        init() {
-            const sync = () => {
-                const url = new URL(window.location.href);
-                const set = (key, value, defaultValue) => {
-                    if (value && value !== defaultValue) {
-                        url.searchParams.set(key, value);
-                    } else {
-                        url.searchParams.delete(key);
-                    }
-                };
-                set('mode', this.viewMode, 'tactical');
-                set('pos', this.posFilter, 'all');
-                set('avail', this.availFilter, 'all');
-                set('sort', this.sortCol, null);
-                set('dir', this.sortDir, 'desc');
-                history.replaceState({}, '', url);
-            };
-            this.$watch('viewMode', sync);
-            this.$watch('posFilter', sync);
-            this.$watch('availFilter', sync);
-            this.$watch('sortCol', () => { this.applySort(); sync(); });
-            this.$watch('sortDir', () => { this.applySort(); sync(); });
-            sync();
-            this.applySort();
-        },
-        // Text columns sort ascending first; numeric columns sort descending first.
-        textCols: ['name', 'pos'],
-        toggleSort(col) {
-            if (this.sortCol === col) {
-                const firstDir = this.textCols.includes(col) ? 'asc' : 'desc';
-                if (this.sortDir === firstDir) {
-                    // Second click: flip direction.
-                    this.sortDir = firstDir === 'asc' ? 'desc' : 'asc';
-                } else {
-                    // Third click: restore default (position-then-rating) order.
-                    this.sortCol = null;
-                    this.sortDir = 'desc';
-                }
-            } else {
-                this.sortCol = col;
-                this.sortDir = this.textCols.includes(col) ? 'asc' : 'desc';
-            }
-        },
-        applySort() {
-            const rows = this.$refs.rows ? Array.from(this.$refs.rows.querySelectorAll('.player-row')) : [];
-            if (!this.sortCol) {
-                rows.forEach((el) => { el.style.order = ''; });
-                return;
-            }
-            const col = this.sortCol;
-            const isText = this.textCols.includes(col);
-            const dir = this.sortDir === 'asc' ? 1 : -1;
-            const sorted = rows.slice().sort((a, b) => {
-                const av = a.dataset['sort' + col.charAt(0).toUpperCase() + col.slice(1)] ?? '';
-                const bv = b.dataset['sort' + col.charAt(0).toUpperCase() + col.slice(1)] ?? '';
-                if (isText) return dir * String(av).localeCompare(String(bv));
-                return dir * ((parseFloat(av) || 0) - (parseFloat(bv) || 0));
-            });
-            sorted.forEach((el, i) => { el.style.order = i; });
-        },
-        isVisible(group, available, status) {
-            if (this.posFilter !== 'all' && group !== this.posFilter) return false;
-            if (this.availFilter === 'available' && !available) return false;
-            if (this.availFilter === 'unavailable' && available) return false;
-            if (this.statusFilter !== 'all' && status !== this.statusFilter) return false;
-            return true;
-        },
-        activeFilterCount() {
-            let c = 0;
-            if (this.posFilter !== 'all') c++;
-            if (this.availFilter !== 'all') c++;
-            if (this.statusFilter !== 'all') c++;
-            return c;
-        },
-        clearFilters() {
-            this.posFilter = 'all';
-            this.availFilter = 'all';
-            this.statusFilter = 'all';
-        }
-    }">
+    <div x-data="squadOverview()">
         <div class="max-w-7xl mx-auto px-4 pb-8">
 
             {{-- Sub-navigation --}}
@@ -203,15 +115,16 @@
                             <div class="grid items-center px-4 py-2 bg-surface-700/30 border-b border-border-default text-[10px] text-text-muted uppercase tracking-widest font-semibold"
                                  :class="{
                                     @if($isCareerMode)
-                                    'grid-cols-[1fr_48px_32px_52px_88px_88px_64px_64px_56px] gap-1.5': viewMode === 'tactical',
+                                    'grid-cols-[36px_248px_76px_52px_52px_80px_64px_84px_80px_80px_1fr] gap-1.5': viewMode === 'tactical',
                                     @else
-                                    'grid-cols-[1fr_48px_32px_52px_88px_88px] gap-1.5': viewMode === 'tactical',
+                                    'grid-cols-[36px_248px_76px_52px_52px_80px_64px_1fr] gap-1.5': viewMode === 'tactical',
                                     @endif
-                                    'grid-cols-[1fr_48px_32px_52px_88px_64px_64px_56px_80px] gap-1.5': viewMode === 'planning',
-                                    'grid-cols-[1fr_48px_32px_52px_48px_48px_48px_40px_48px_48px_48px_64px] gap-1.5': viewMode === 'stats',
+                                    'grid-cols-[36px_248px_76px_52px_52px_112px_84px_80px_80px_1fr] gap-1.5': viewMode === 'planning',
+                                    'grid-cols-[36px_248px_76px_52px_52px_48px_44px_44px_40px_44px_48px_44px_72px_1fr] gap-1.5': viewMode === 'stats',
                                  }">
+                                <x-squad.sort-header col="number">#</x-squad.sort-header>
                                 <x-squad.sort-header col="name" align="left">{{ __('squad.player') }}</x-squad.sort-header>
-                                <x-squad.sort-header col="pos">{{ __('squad.pos') }}</x-squad.sort-header>
+                                <x-squad.sort-header col="pos" align="left">{{ __('squad.pos') }}</x-squad.sort-header>
                                 <x-squad.sort-header col="age">{{ __('app.age') }}</x-squad.sort-header>
                                 <x-squad.sort-header col="ovr">{{ __('transfers.explore_overall') }}</x-squad.sort-header>
 
@@ -247,9 +160,6 @@
                                 </template>
                                 <template x-if="viewMode === 'planning'">
                                     <x-squad.sort-header col="contract">{{ __('app.contract') }}</x-squad.sort-header>
-                                </template>
-                                <template x-if="viewMode === 'planning'">
-                                    <x-squad.sort-header col="dev">{{ __('squad.dev_status_label') }}</x-squad.sort-header>
                                 </template>
                                 @endif
 
@@ -289,13 +199,12 @@
                                 $isUnavailable = $gp->is_unavailable;
                                 $unavailReason = $gp->unavailability_reason;
                                 $groupKey = $gp->position_group;
-                                $posAbbrev = \App\Support\PositionMapper::toAbbreviation($gp->position);
+                                // Sort key: canonical position order (GK → defenders → midfielders →
+                                // forwards, with specific positions ordered within each line). The stable
+                                // client-side sort keeps the default rating-desc order among players who
+                                // share the same position.
+                                $posOrder = \App\Support\PositionMapper::positionSortOrder($gp->position);
                                 $mvpCount = $mvpCounts[$gp->id] ?? 0;
-                                $devOrdinal = match ($gp->dev_status) {
-                                    'growing' => 2,
-                                    'peak' => 1,
-                                    default => 0,
-                                };
 
                                 $statusKey = 'none';
                                 if ($isCareerMode) {
@@ -309,7 +218,8 @@
                             <div x-show="isVisible('{{ $groupKey }}', {{ $isUnavailable ? 'false' : 'true' }}, '{{ $statusKey }}')"
                                  class="player-row border-b border-border-default {{ $isUnavailable ? 'opacity-60' : '' }}"
                                  data-sort-name="{{ \Illuminate\Support\Str::lower($gp->name) }}"
-                                 data-sort-pos="{{ $posAbbrev }}"
+                                 data-sort-number="{{ $gp->number ?? 999 }}"
+                                 data-sort-pos="{{ $posOrder }}"
                                  data-sort-age="{{ $gp->age($game->current_date) }}"
                                  data-sort-ovr="{{ $gp->effective_rating }}"
                                  data-sort-fitness="{{ $gp->fitness }}"
@@ -318,7 +228,6 @@
                                  data-sort-wage="{{ $gp->annual_wage }}"
                                  data-sort-contract="{{ $gp->contract_expiry_year ?? 0 }}"
                                  data-sort-potential="{{ $gp->potential_high }}"
-                                 data-sort-dev="{{ $devOrdinal }}"
                                  data-sort-apps="{{ $gp->appearances }}"
                                  data-sort-goals="{{ $gp->goals }}"
                                  data-sort-assists="{{ $gp->assists }}"
@@ -395,21 +304,23 @@
                                     </div>
 
                                     {{-- ===== DESKTOP ROW ===== --}}
-                                    <div class="hidden lg:grid items-center px-4 py-2.5 gap-2 cursor-pointer"
+                                    <div class="hidden lg:grid items-center px-4 py-2.5 cursor-pointer"
                                          @click="$dispatch('show-player-detail', '{{ route('game.player.detail', [$game->id, $gp->id]) }}')"
                                          :class="{
                                             @if($isCareerMode)
-                                            'grid-cols-[1fr_48px_32px_52px_88px_88px_64px_64px_56px] gap-1.5': viewMode === 'tactical',
+                                            'grid-cols-[36px_248px_76px_52px_52px_80px_64px_84px_80px_80px_1fr] gap-1.5': viewMode === 'tactical',
                                             @else
-                                            'grid-cols-[1fr_48px_32px_52px_88px_88px] gap-1.5': viewMode === 'tactical',
+                                            'grid-cols-[36px_248px_76px_52px_52px_80px_64px_1fr] gap-1.5': viewMode === 'tactical',
                                             @endif
-                                            'grid-cols-[1fr_48px_32px_52px_88px_64px_64px_56px_80px] gap-1.5': viewMode === 'planning',
-                                            'grid-cols-[1fr_48px_32px_52px_48px_48px_48px_40px_48px_48px_48px_64px] gap-1.5': viewMode === 'stats',
+                                            'grid-cols-[36px_248px_76px_52px_52px_112px_84px_80px_80px_1fr] gap-1.5': viewMode === 'planning',
+                                            'grid-cols-[36px_248px_76px_52px_52px_48px_44px_44px_40px_44px_48px_44px_72px_1fr] gap-1.5': viewMode === 'stats',
                                          }">
 
-                                        {{-- Player name with avatar --}}
-                                        <div class="flex items-center gap-3 min-w-0">
-                                            <x-player-avatar :name="$gp->name" :position-group="$groupKey" :number="$gp->number" size="sm" />
+                                        {{-- Squad number --}}
+                                        <span class="text-center font-heading tabular-nums text-sm text-text-muted">{{ $gp->number }}</span>
+
+                                        {{-- Player: name + nationality --}}
+                                        <div class="flex items-center gap-2.5 min-w-0">
                                             <div class="min-w-0">
                                                 <div class="flex items-center gap-2">
                                                     <span class="text-sm font-medium text-text-primary truncate">{{ $gp->name }}</span>
@@ -435,7 +346,7 @@
                                         </div>
 
                                         {{-- Position badges --}}
-                                        <div class="flex items-center justify-center gap-0.5">
+                                        <div class="flex items-center justify-start gap-0.5">
                                             @foreach($gp->positions as $pos)
                                                 <x-position-badge :position="$pos" size="sm" />
                                             @endforeach
@@ -451,10 +362,10 @@
 
                                         {{-- === Tactical columns === --}}
                                         <template x-if="viewMode === 'tactical'">
-                                            <x-fitness-bar :value="$gp->fitness" class="justify-center" />
+                                            <x-fitness-bar :value="$gp->fitness" size="sm" class="justify-center" />
                                         </template>
                                         <template x-if="viewMode === 'tactical'">
-                                            <x-morale-indicator :value="$gp->morale" class="justify-center" />
+                                            <x-morale-indicator :value="$gp->morale" :show-label="false" class="justify-center" />
                                         </template>
                                         @if($isCareerMode)
                                         <template x-if="viewMode === 'tactical'">
@@ -476,6 +387,7 @@
                                                     :current-ability="$gp->overall_score"
                                                     :potential-low="$gp->potential_low"
                                                     :potential-high="$gp->potential_high"
+                                                    :projection="$gp->projection"
                                                     size="sm"
                                                 />
                                             </div>
@@ -490,20 +402,6 @@
                                             <span class="text-[11px] text-center tabular-nums @if($gp->isContractExpiring($seasonEndDate)) text-accent-red font-medium @else text-text-muted @endif">
                                                 {{ $gp->contract_expiry_year ?? '' }}
                                             </span>
-                                        </template>
-                                        <template x-if="viewMode === 'planning'">
-                                            <div class="flex items-center justify-center">
-                                                @php $ds = $gp->dev_status; @endphp
-                                                <span class="shrink-0 @if($ds === 'growing') text-accent-green @elseif($ds === 'peak') text-accent-blue @else text-orange-600 @endif" x-data x-tooltip.raw="{{ __('squad.' . $ds) }}">
-                                                    @if($ds === 'growing')
-                                                        <svg class="w-3.5 h-3.5" fill="none" stroke="currentColor" stroke-width="3" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M5 15l7-7 7 7"/></svg>
-                                                    @elseif($ds === 'declining')
-                                                        <svg class="w-3.5 h-3.5" fill="none" stroke="currentColor" stroke-width="3" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M19 9l-7 7-7-7"/></svg>
-                                                    @else
-                                                        <svg class="w-3.5 h-3.5" fill="none" stroke="currentColor" stroke-width="3" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M5 12h14"/></svg>
-                                                    @endif
-                                                </span>
-                                            </div>
                                         </template>
                                         @endif
